+++ packages/grafana-ui/src/utils/ValueFormats/symbolFormatters.test.ts
@@ -1,0 +1,11 @@
+import { currency } from './symbolFormatters';
+
+describe('Currency', () => {
+  it('should format as usd', () => {
+    expect(currency('$')(1532.82, 1, -1)).toEqual('$1.53K');
+  });
+
+  it('should format as usd', () => {
+    expect(currency('kr')(1532.82, 1, -1)).toEqual('1.53K kr');
+  });
+});
